Fixed command

+++ src/main/java/bronuh/shit/metrics/Stat.java
@@ -2,9 +2,9 @@ package bronuh.shit.metrics;
 public class Stat {
     public String statId;
     public String statName;
-    public int value;
+    public double value;
 
-    public Stat(String statId, String statName, int value){
+    public Stat(String statId, String statName, double value){
         this.statId = statId;
         this.statName = statName;
         this.value = value;
